% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{create_data}
\usage{
create_data(
  n,
  p_boost,
  boost_assignment = c("random", "vulnerable"),
  post_boost_disinhibition = c(0, 1),
  lambda_0_N,
  lambda_0_Y,
  init_VE,
  VE_wane,
  Tmax,
  sd_frail
)
}
\arguments{
\item{n}{numeric, number of participants to simulate data for}

\item{p_boost}{numeric, probability of vaccination at some point during study}

\item{boost_assignment}{character, "random" if vaccination date is not associated with risk, "vulnerable" if to prioritize vaccinating at-risk persons early}

\item{post_boost_disinhibition}{binary, 0 if no increase in risk post-boost, 1 if increase in risk post-boost}

\item{lambda_0_N}{numeric, time-averaged baseline hazard of vaccine-irrelevant infection}

\item{lambda_0_Y}{numeric, time-averaged baseline hazard of vaccine-preventable infection}

\item{init_VE}{numeric, log(HR) of infection just after vaccination}

\item{VE_wane}{numeric, how log(HR) varies in time-since-vaccination}

\item{Tmax}{numeric, duration of study}

\item{sd_frail}{numeric, level of risk heterogeneity, defined as standard deviation of the lognormal frailty variable}
}
\value{
list containing elements "covars" with summary info (including id, frailties, and vaccination status), "dat_N" containing vaccine-irrelevant infetcion data, "dat_Y_const" containing vaccine-preventable infection data with constant VE, and "dat_Y_wane" containing vaccine-preventable infection data with waning VE.
}
\description{
Function used to generate survival data consistent with RCTs/observational studies measuring vaccine-preventable and irrelevant infections/
}
\details{
The following code generates data under two scenarios (i) a constant VE setting and (ii) a linearly waning VE setting on the log hazard ratio scale.
Survival outcomes (for both vaccine-preventable and vaccine-irrelevant pathogens) are generated from hazard frailty models.
}
\examples{
create_data(n=100, p_boost=0.70, boost_assignment="vulnerable", post_boost_disinhibition=1, lambda_0_N=0.06, lambda_0_Y=0.08, init_VE=1.3, VE_wane = -0.5, Tmax=1, sd_frail=1.3)
}
