% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_f.R
\name{linear_basis}
\alias{linear_basis}
\title{linear_basis}
\usage{
linear_basis(x)
}
\arguments{
\item{x}{Vector of time points}
}
\value{
An object of type 'linear_basis'
}
\description{
Helper function to create linear basis
}
