% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sieve_multinomial.r
\name{partially_linear_multinomial_sieve}
\alias{partially_linear_multinomial_sieve}
\title{partially_linear_multinomial_sieve}
\usage{
partially_linear_multinomial_sieve(
  dat,
  J.name = "J",
  V.name = "V",
  V.early.name = NULL,
  T.name = "T",
  S.name = "S",
  m,
  psi_delta,
  p_s_fun = NULL,
  verbose = TRUE,
  tol = 1e-06,
  calc_mixture = FALSE,
  ...
)
}
\arguments{
\item{dat}{data.frame, containing columns J.name, V.name, T.name, and V.early.name (optional).}

\item{J.name}{character, column name associated with outcome type (J=0 vaccine-irrelevant, J=1 vaccine-preventable)}

\item{V.name}{character, column name associated with vaccination date or peak immunization date (if V.early.name is specified)}

\item{V.early.name}{character, OPTIONAL column name associated with "early" effectiveness (often the vaccination date itself)}

\item{T.name}{character, column name associated with infection date}

\item{S.name}{character, column name associated with the strain name (S=NA for irrelevant, S=1 for strain 1, S=2 for strain 2, etc.)}

\item{m}{integer, number of strains of vaccine-preventable infection in the dataset}

\item{psi_delta}{function, used to build the basis for time-varying VE}

\item{p_s_fun}{function, OPTIONAL function that takes as input a vector of times and outputs mixture probabilities for the different strains}

\item{verbose}{logical, whether to print out intermediate info}

\item{tol}{numeric, tolerance limit to stop iterations}

\item{calc_mixture}{logical, whether or not to calculate variant mixture proportions from the data itself or rely on external source}

\item{...}{extra arguments to use for basis creation functions}
}
\value{
list containing parameter estimates associated with VE, SEs, vcov matrix, the varying intercept alpha(t), and the basis
}
\description{
Function used to estimate time-varying VE using the partially linear \emph{multinomial} logistic regression model adjusting for vaccine-irrelevant infections via the sieve estimation method.
}
\details{
The code has an option to profile out nuisance parameters (the variant mixture proportions) by using external disease surveillance data.
}
