% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_TMLE.r
\name{partially_linear_multinomial_tmle}
\alias{partially_linear_multinomial_tmle}
\title{partially_linear_multinomial_tmle}
\usage{
partially_linear_multinomial_tmle(
  dat,
  T.name = "T",
  J.name = "J",
  V.name = "V",
  V.early.name = "V_early",
  S.name = "S",
  psi_delta,
  m = 2,
  p_s_fun = NULL,
  maxit = 100,
  tol = 1e-06,
  kernel_shape = c("gaussian", "epanechikov"),
  bandwidth = NULL,
  automatic_bw = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{data.frame, containing columns J.name, V.name, T.name, and V.early.name (optional).}

\item{T.name}{character, column name associated with infection date}

\item{J.name}{character, column name associated with outcome type (J=0 vaccine-irrelevant, J=1 vaccine-preventable)}

\item{V.name}{character, column name associated with vaccination date or peak immunization date (if V.early.name is specified)}

\item{V.early.name}{character, OPTIONAL column name associated with "early" effectiveness (often the vaccination date itself)}

\item{S.name}{character, column name associated with the strain name (S=NA for irrelevant, S=1 for strain 1, S=2 for strain 2, etc.)}

\item{psi_delta}{function, used to build the basis for time-varying VE}

\item{m}{integer, number of strains of vaccine-preventable infection in the dataset}

\item{p_s_fun}{function, OPTIONAL function that takes as input a vector of times and outputs mixture probabilities for the different strains}

\item{maxit}{integer, maximum number of TMLE iterations}

\item{tol}{numeric, tolerance limit to stop iterations}

\item{kernel_shape}{character, kernel shape for nuisance parameter estimation. Options are "gaussian" and "epanechikov"}

\item{bandwidth, }{numeric, OPTIONAL, the bandwidth for the kernel estimator for nuisance parameters}

\item{automatic_bw}{logical, whether or not to use LOOCV to choose bandwidth. If FALSE and bandwidth is not provided, will use Silverman's rule of thumb.}

\item{verbose}{logical, whether to print out intermediate info}

\item{...}{extra arguments to use for basis creation functions}
}
\value{
list containing parameter estimates associated with VE, SEs, cov matrix, the varying intercept alpha(t), and the basis
}
\description{
Function used to estimate time-varying VE using the partially linear \emph{multinomial} logistic regression model adjusting for vaccine-irrelevant infections via the iterative TMLE method.
}
\details{
The code has an option to profile out nuisance parameters (the variant mixture proportions) by using external disease surveillance data.
}
