% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_f.R
\name{predict.linear_basis}
\alias{predict.linear_basis}
\title{predict.linear.basis}
\usage{
\method{predict}{linear_basis}(object, newx, ...)
}
\arguments{
\item{object}{Object of type 'linear_basis'}

\item{newx}{Vector of time points you want predictions}

\item{...}{Other arguments to the predict method (shouldn't need to be used)}
}
\value{
Matrix of predicted effectiveness
}
\description{
Predict method for linear basis
}
