% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_f.R
\name{psi_bs_early}
\alias{psi_bs_early}
\title{psi_bs_early}
\usage{
psi_bs_early(
  Tvec,
  Vvec,
  Vvec_early,
  df,
  left_boundary_knot,
  leftmost_interior_knot,
  intercept
)
}
\arguments{
\item{Tvec}{Vector of infection dates}

\item{Vvec}{Vector of dates at which vaccine is presumed to achieve peak effectiveness (often vaccination date + 14 days)}

\item{Vvec_early}{Vector of \emph{actual} vaccination dates}

\item{df}{Degrees of freedom of splines}

\item{left_boundary_knot}{Location of left boundary knot (to control left boundary behavior)}

\item{leftmost_interior_knot}{Location of leftmost interior knot}

\item{intercept}{Logical, whether or not to include intercept in your basis}
}
\value{
A list containing (i) a basis matrix evaluated at the inputs Tvec and Vvec, and (ii) the 'linear_basis' object for which additional predictions can be obtained
}
\description{
Create cubic b-spline basis (with additional scalar dimension for "early" vaccination)
}
