% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_f.R
\name{psi_d2}
\alias{psi_d2}
\title{psi_d2}
\usage{
psi_d2(Tvec, Vvec)
}
\arguments{
\item{Tvec}{Vector of infection dates}

\item{Vvec}{Vector of vaccination dates}
}
\value{
A list containing (i) a basis matrix evaluated at the inputs Tvec and Vvec, and (ii) the 'linear_basis' object for which additional predictions can be obtained
}
\description{
Create two dimensional linear basis
}
