% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_f.R
\name{psi_d2_early}
\alias{psi_d2_early}
\title{psi_d2_early}
\usage{
psi_d2_early(Tvec, Vvec, Vvec_early)
}
\arguments{
\item{Tvec}{Vector of infection dates}

\item{Vvec}{Vector of dates at which vaccine is presumed to achieve peak effectiveness (often vaccination date + 14 days)}

\item{Vvec_early}{Vector of \emph{actual} vaccination dates}
}
\value{
A list containing (i) a basis matrix evaluated at the inputs Tvec and Vvec, and (ii) the 'linear_basis' object for which additional predictions can be obtained
}
\description{
Create three dimensional linear basis (with additional scalar dimension for "early" vaccination)
}
