% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sieve_func.R
\name{sieve_partially_linear_logistic}
\alias{sieve_partially_linear_logistic}
\title{sieve_partially_linear_logistic}
\usage{
sieve_partially_linear_logistic(
  dat,
  J.name = "J",
  V.name = "V",
  V.early.name = NULL,
  T.name = "T",
  psi_delta,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{data.frame, containing columns J.name, V.name, T.name, and V.early.name (optional).}

\item{J.name}{character, column name associated with outcome type (J=0 vaccine-irrelevant, J=1 vaccine-preventable)}

\item{V.name}{character, column name associated with vaccination date or peak immunization date (if V.early.name is specified)}

\item{V.early.name}{character, OPTIONAL column name associated with "early" effectiveness (often the vaccination date itself)}

\item{T.name}{character, column name associated with infection date}

\item{psi_delta}{function, used to build the basis for time-varying VE}

\item{verbose}{logical, whether to print out intermediate info}

\item{...}{extra arguments to use for basis creation functions}
}
\value{
list containing parameter estimates associated with VE, SEs, vcov matrix, the varying intercept alpha(t), and the basis
}
\description{
Function used to estimate time-varying VE using the partially linear logistic regression model adjusting for vaccine-irrelevant infections via the sieve method.
}
