% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeVaryingVE-package.R
\docType{package}
\name{TimeVaryingVE}
\alias{TimeVaryingVE-package}
\alias{TimeVaryingVE}
\title{TimeVaryingVE: Tools for Time-Varying Vaccine Effectiveness}
\description{
This package includes R functions to estimate time-varying vaccine effectiveness adjusting for vaccine-irrelevant infections to avoid bias.
}
\author{
\strong{Maintainer}: Ethan Ashby \email{ecadeashby@gmail.com}

}
\keyword{internal}
