% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMLE_func.R
\name{tmle_partially_linear_logistic}
\alias{tmle_partially_linear_logistic}
\title{tmle_partially_linear_logistic}
\usage{
tmle_partially_linear_logistic(
  dat,
  J.name = "J",
  T.name = "T",
  V.early.name = NULL,
  V.name = "V",
  psi_delta,
  maxit = 500,
  tol = 1e-06,
  smooth_r = TRUE,
  smooth_alpha = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{data.frame, containing columns J.name, V.name, T.name, and V.early.name (optional).}

\item{J.name}{character, column name associated with outcome type (J=0 vaccine-irrelevant, J=1 vaccine-preventable)}

\item{T.name}{character, column name associated with infection date}

\item{V.early.name}{character, OPTIONAL column name associated with "early" effectiveness (often the vaccination date itself)}

\item{V.name}{character, column name associated with vaccination date or peak immunization date (if V.early.name is specified)}

\item{psi_delta}{function, used to build the basis for time-varying VE}

\item{maxit}{integer, maximum number of TMLE iterations}

\item{tol}{numeric, tolerance limit to stop iterations}

\item{smooth_r}{logical, for whether to compute the nuisance parameter via kernel smoothing}

\item{smooth_alpha}{logical, for whether to smooth out the estimate of logistic model time-varying intercept between iterations.}

\item{verbose}{logical, whether to print out intermediate info}

\item{...}{extra arguments to use for basis creation functions}
}
\value{
list containing parameter estimates associated with VE, SEs, vcov matrix, the varying intercept alpha(t), and the basis
}
\description{
Function used to estimate time-varying VE using the partially linear logistic regression model adjusting for vaccine-irrelevant infections via the iterative TMLE method.
}
